// frontend/ReceiptTracker/src/app/api/model/publicUserRegister.ts
/**
 * Receipt Tracker API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface PublicUserRegister { 
    username: string;
    email?: string | null;
    fullname?: string | null;
    password: string;
}